change loadScript suffix, move to lib folder, remove unused SurveyModal
